Implement RootType on failfs
A RootType method was added to the FileSystem interface in order to
indicate whether the FS represents a GOROOT or a GOPATH. In the case
that the FS represents neither, an empty string is used.

The method was added to the interface in
16d1af8d88c08b6a22e07a0900aad5279a7fb4fd in github.com/golang/tools, and
the implementation was updated in
8b3cccae50961d427273d41c8a6ad18bfa623e08 to clarify the usage of empty
string as the default value.